requirements.txt
aniso8601==1.0.0
bcrypt==2.0.0
factory-boy==2.7.0
fake-factory==0.5.7
Flask==0.11.0
flask-cors==2.1.2
Flask-API==0.6.5
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.4.2
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
mock==2.0.0
passlib==1.6.5
psycopg2==2.6
pyjwt==1.4.0
python-dateutil==2.5.3
pytz==2015.4
six==1.9.0
SQLAlchemy==1.0.4
titlecase==0.8.1
Werkzeug==0.10.4
webargs==1.3.2

-e git+https://github.com/jjwon0/plaid-python@215a8e289920ecad8d0b075bf0db1ce1a6e3a5bc#egg=plaid_python
-e git+https://github.com/phoxelua/flask-shellplus@04ac9a9349a944311a4b6615ebf41713870f74e1#egg=flask-shellplus
